requirements.txt
Package                       Version            
----------------------------- -------------------
gensim                        4.0.1              
ginza                         5.1.0              
ginza-transformers            0.4.0              
ipadic                        1.0.0              
ja-ginza                      5.1.0              
ja-ginza-dict                 3.1.0              
Janome                        0.4.1              
jupyter-book                  0.8.3              
jupyter-cache                 0.4.1              
jupyter-client                6.1.7              
jupyter-core                  4.7.0              
jupyter-server                1.2.1              
jupyter-sphinx                0.3.1              
jupyterlab                    3.0.3              
jupyterlab-server             2.1.1              
jupytext                      1.11.5             
mecab-python3                 1.0.4              
networkx                      2.5                
openpyxl                      3.0.5              
Pillow                        8.0.1              
prince                        0.7.1              
regex                         2021.4.4           
scikit-learn                  0.23.2             
scipy                         1.5.4              
seaborn                       0.11.1             
sentencepiece                 0.1.95             
spacy                         3.2.1              
spacy-alignments              0.8.4              
spacy-legacy                  3.0.8              
spacy-loggers                 1.0.1              
spacy-transformers            1.1.3              
transformers                  4.12.5             
wordcloud                     1.8.1              
